added launcher, fixed crash on hitting next track after sample stopped.

+++ pyomx.py
@@ -35,9 +35,10 @@ class Player(object):
 
     def stop(self):
         print("stop")
-        self.paused = True
-        self.process.stdin.write(self.STOP)
-        self.process.terminate()
+        if (self.paused == False):
+          self.paused = True
+          self.process.stdin.write(self.STOP)
+          self.process.terminate()
 
     def pollDone(self):
        code = self.process.poll()
